Author: Amir Reviewer: JP
This branch does 2 things: keeps the stats more lively and increments
the public click count aggressively.

For the lively stats, a record is kept of the lowest score of the top
25 scores. If a stat is updated that puts it into the top 25, the rest
of the top are recalculated. This was necessary because an url could
become popular and never receive another click. This would keep it at
the top of popular indefinitely. With this new code, it will be
recalculated and fall. Wierd behaviours are still possible, but less
likely.

For the aggressive count, I've hacked the aggregation protocol to use
a special token as the url when all we want to do is increment the
count which is the case in the following 2 scenarios: a user clicks on
an url that they've already posted to the public page in the last 24
hours or the user has their anonymous sharing disabled.



git-svn-id: http://divmod.org/svn/Divmod/trunk@3004 866e43f7-fbfc-0310-8f2a-ec88d1da2979