Fixes #327
This test is just wrong.  It served its purpose, which was to locate a silly
Axiom->Mantissa dependency, but it's broken, and should be replaced.  Phil
wrote a test that checks this which is in old Quotient in
quotient.test.test_import or so.

Now we can run Epsilon and Axiom tests together.


git-svn-id: http://divmod.org/svn/Divmod/trunk@3085 866e43f7-fbfc-0310-8f2a-ec88d1da2979