Make it possible to extract users from a store and insert them into a new store.
Author: glyph+exarkun
Reviewer: dash

Fixes #409

This adds 'axiomatic extract' and 'axiomatic deploy' subcommands which can
extract and insert users into the store, thereby this unblocking re-deployment
of clickchronicle.

This branch also includes some other fixes, such as a 'hotfix' system, the
first 'hotfix' in that system (the new functionality added in
twisted.python.filepath in Twisted SVN-Trunk), more consistent treatment of the
'store' argument to infilter so that e.g. timestamps can be passed to
initializers, and more portable relative path handling in the database (so you
can make stores on Windows and move them to Linux if they contain
store-relative filenames).


git-svn-id: http://divmod.org/svn/Divmod/trunk@3468 866e43f7-fbfc-0310-8f2a-ec88d1da2979