Merge user-browser-261
Author: exarkun
Reviewer: moe
Fixes #261

This adds an INavigableElement to the administrative benefactor.  
When rendered, it displays a TDB of all non-system users in the
local authentication database.  Users with multiple login methods
are listed multiple times.

The mantissa admin-creation code has also been changed to use the
administrative benefactor rather than duplicating the list of
powerups it contains.


git-svn-id: http://divmod.org/svn/Divmod/trunk@3516 866e43f7-fbfc-0310-8f2a-ec88d1da2979