Axiom databases will no longer open if doing so will corrupt them.
Author: glyph

Reviewer: moe

Fixes #301

Previously a database with missing intermediary schema would happilly open and
destroy data.  Now, it will first verify that all old versions are available
for full upgrading.

Special thanks to Moe for finding some old, broken production data which would
not load with this branch and testing it, as well as a good suggestion on error
messages.


git-svn-id: http://divmod.org/svn/Divmod/trunk@4599 866e43f7-fbfc-0310-8f2a-ec88d1da2979