Merge faster-grabber-630
Author: exarkun
Reviewer: glyph
Fixes #630

Instead of building up a list of all UIDs in the POP3 mailbox and then filtering
them to find the ones which should be retrieved, do the filtering as the UID list
is streamed to the client.  This avoids keeping a potentially huge list of UIDs
in memory some of the time (when they have already been downloaded), and spreads
the work of filtering out the UIDs which must be downloaded out over the course
of receiving the UIDL response.  This removes the need to periodically yield to
the reactor while downloading messages, since we will already be yielding on
every message, rather than only on some messages as previously was the case.  This
also removes the large delay between receiving the last entry of the UIDL response
and issuing the first RETR command.


git-svn-id: http://divmod.org/svn/Divmod/trunk@4636 866e43f7-fbfc-0310-8f2a-ec88d1da2979