Properly update download status - fixes bug introduced by faster-grabber-630
Previously, uidList contained all UIDs downloaded from the server, making len(uidList)
a fairly accurate indicator of total work to be done.  However, the #630 branch changed
this so that uidList only held UIDs which were going to be downloaded.  This changes the
reporting code to use the last index reported by the server in the UIDL response as the
total number of messages when generating a status message.


git-svn-id: http://divmod.org/svn/Divmod/trunk@4667 866e43f7-fbfc-0310-8f2a-ec88d1da2979