Merge isometric-682
Author: exarkun
Reviewer: moe (Mantissa bits)
Fixes #682

This adds the beginning of an Athena-based Radical implementation.  It also
adds /live/ to Mantissa to short-circuit all the unnecessary database work
when serving a LivePage's transports, and makes it even more convenient 
to use LiveFragment by removing the need to set `live' on it when it
is used as an INavigableFragment.


git-svn-id: http://divmod.org/svn/Divmod/trunk@4850 866e43f7-fbfc-0310-8f2a-ec88d1da2979