More conservative transaction management.
BEGIN IMMEDIATE TRANSACTION when transactions start. Remove implicit management
of transactions, replace with real autocommit. As a bonus, optionally support
APSW - I'm concerned that PySQLite's transaction handling will not help us
here.  APSW is considerably more straightforward here, and wraps one or two
additional features from sqlite.  However, APSW's unicode handling is
definitely wrong - it attempts to return bytestrings when stuff is ASCII (or
maybe latin-1??), so it's not a clear win.

The fact that the transactions are IMMEDIATE means that database access will be
_fully_ serialized, at least with respect to transactions.  I should do a
little more reading if contention on the database becomes problematic - but
reading from the database should not cause issues, as it is not EXCLUSIVE.

Still, the two APIs are remarkably similar.

Fixes #659

Author: glyph
Reviewer: exarkun


git-svn-id: http://divmod.org/svn/Divmod/trunk@4869 866e43f7-fbfc-0310-8f2a-ec88d1da2979