Merge indexer-subprocess-655-3
Author: washort, exarkun
Reviewer: moe

This adds a child process, managed by a service automatically associated
on-disk Stores, started when the Store service is started, stopped when it is
stopped, accessible by adapting a Store to axiom.iaxiom.IBatchService, which
processes certain kinds of batch tasks.

Process management code is factored out of Sine and into a generally usable set
of classes provided by Axiom.

The hype indexer is changed to declare its work as out-of-process.  Messages
are passed between the two processes to coordinate searches, due to the
constraints hype places on concurrent access to the index.

Historical unit tests are cleaned up to use setUp and tearDown properly, so as
not to start services but never stop them.  Additionally, the code is now
essentially identical across all of the historical tests, making it extremely
ripe for factoring out.

The API for reliable listeners is reified as axiom.iaxiom.IReliableListener and
expanded to support suspend and resume messages.

The search interface defined by mantissa is now asynchronous, requiring
Deferred returns from count and search.  ClickChronicle's xapian
search-provider has been updated to comply with this.

Several debugging aids useful during the course of development of this branch
have been added.  epsilon.spewer is a prettier version of
twisted.python.util.spewer.  JuiceBoxes have a prettier repr() string.


git-svn-id: http://divmod.org/svn/Divmod/trunk@5124 866e43f7-fbfc-0310-8f2a-ec88d1da2979