Merge better-concurrency-658
Author: exarkun
Reviewer: washort
Fixes #658

This adds support for two operations possible when concurrently accessing
the same Axiom database from two different threads of control.  The first is
the insertion of the first instance of an Item subclass without causing 
an exception to be raised in the other thread of control.  The second is
the detection /at all/ of instances of a particular Item type if the first 
one was only inserted into the database by one thread of control after another 
thread of control has already opened the database.

This also factors some of the APSW- and PySQLite2-specific code into separate 
modules.


git-svn-id: http://divmod.org/svn/Divmod/trunk@5188 866e43f7-fbfc-0310-8f2a-ec88d1da2979