Import nodes into the current document to fix IMG support on FF10
This one-line fix might work around some other oddities as well, as this is
probably the Way You Should Do It in DOM.

set/appendNodeContent is rapidly becoming the One True Way to manipulate DOM
within Athena.  Keep reporting these bugs - we need to make sure we have a
central way of dealing with issues like this so we don't keep running into
them.

We actually installed Firefox 1.0 to test this.  Can you feel the love, moe?

Author: exarkun

Reviewer: glyph

Fixes #746


git-svn-id: http://divmod.org/svn/Divmod/trunk@5235 866e43f7-fbfc-0310-8f2a-ec88d1da2979