Fixes #624
This does not satisfy the full description on 624, but it turns out that the
full description is flawed.  As stated, such a trash system would have been a
stack; you would be unable (for example) to have a message A,
message-person-link (A->C) B, and person C, delete A, delete B, then undelete
A.  It's not clear what the right thing to do in that situation is in any case,
actually.

Instead, this commit lays the groundwork for regular deletion.  There were
several bugs related to transactionality of deleteFromStore previously, which
have been fixed.  In addition to fixing what was there before, deleteFromStore
will now respect whenDeleted=reference.CASCADE, which causes the item that sets
it to be deleted when the item referred to through the annotated reference is
deleted.

Also, it includes a cross-store transaction layer, and the ability to attach a
substore to its parent for true cross-store transactionality.


git-svn-id: http://divmod.org/svn/Divmod/trunk@5493 866e43f7-fbfc-0310-8f2a-ec88d1da2979