Merge close-it-787
Author: exarkun
Reviewer: glyph
Fixes #787

This fixes the server-side close message so that it is formatted correctly
and changes the client-side code which handles this message to actually
shutdown the transport.  The combined effect is that when an Athena LivePage
is closed, it is actually closed, and the browser will have no remaining
open sockets to the server in service of that page.


git-svn-id: http://divmod.org/svn/Divmod/trunk@5559 866e43f7-fbfc-0310-8f2a-ec88d1da2979