Merge initargs-535
Author: exarkun
Reviewer: glyph
Fixes #535

This adds a new callback method to nevow.athena.LiveFragment, getInitialArguments(),
which is invoked during the rendering process and returns a tuple or list of arguments
which will be passed to the client-side widget's __init__ method.  This should allow
the rendering code for many LiveFragments to be simplified, since data can be provided
to them to render when they load, rather than also providing server-side rendering code.

Radical is also updated to take advantage of this feature in the two places it was doing
something similar in an ad hoc fashion.


git-svn-id: http://divmod.org/svn/Divmod/trunk@5675 866e43f7-fbfc-0310-8f2a-ec88d1da2979