Support expected value for `this' when calling JavaScript non-top-level functions
Author: exarkun
Reviewer: harshaw
Fixes #520

This adds an optional "output" Array to namedAny which can be used to 
collected the objects along the way to the actually named object.  It
changes the client-side "call" handler to use this to find the object
on which functions are being called and passes that as `this', or
null for global functions (*not* the global object).


git-svn-id: http://divmod.org/svn/Divmod/trunk@5809 866e43f7-fbfc-0310-8f2a-ec88d1da2979