avoid inconsistent schema in case of catastrophic failure during "CREATE"
Fixes #754

Author: glyph

Reviewer: washort

Sorry for the long merge, Ed.

This puts the creation of tables and indexes into the same transaction as the
creation of the axiom-specific schema rows which annotate the Python-level
schema and such.

It isn't the same patch originally submitted but this one changes a lot fewer
lines.  There are no associated unit tests because this involves a catastrophic
failure case at a particular point: I couldn't figure out a good strategy for
causing the failure there other than adding code to the implementation to
support this one particular test, which seemed counterproductive given the
nature of the feature.


git-svn-id: http://divmod.org/svn/Divmod/trunk@6070 866e43f7-fbfc-0310-8f2a-ec88d1da2979