Merge event-ordering-848
Author: exarkun
Reviewer: radix
Fixes #848, #906


This makes actions run in a transaction and adds a system for making sure
events are displayed to observers in the correct order.

Ordering is made correct by collecting events over the lifetime of the action
transaction and only broadcasting them if the transaction is successfully
committed.  If the transaction is reverted then a separate set of revert
events are broadcast instead.


git-svn-id: http://divmod.org/svn/Divmod/trunk@6267 866e43f7-fbfc-0310-8f2a-ec88d1da2979