Merge batch-shutdown-restart-854
Author: exarkun
Reviewer: moe
Fixes #854

This fixes a couple state bugs in ProcessController which prevented it from
properly invoking the shutdown callback in all cases.  It also changes the
installed shutdown callback to delay slightly between restarts in case there
is some problem which causes fast failures of the batch process.  It also
makes the batch process notice when it has lost its connection to the parent
process and shut itself down.


git-svn-id: http://divmod.org/svn/Divmod/trunk@6352 866e43f7-fbfc-0310-8f2a-ec88d1da2979