Merge repeated-renderer-862
Author: exarkun
Reviewer: moe
Fixes #862

This fixes handling of {{{render_}}} methods which return a {{{Deferred}}} which
ultimately fires with a function.  Previously the function would be invoked more
times than one would expect and more times than is strictly necessary.  It is
now invoked only the necessary number of times (generally once).


git-svn-id: http://divmod.org/svn/Divmod/trunk@6355 866e43f7-fbfc-0310-8f2a-ec88d1da2979