Merge faster-pop3-grabber-1023
Author: exarkun
Reviewer: moea
Fixes #1023

This adds a test for the behavior of xquotient.grabber.POP3Grabber.shouldRetrieve() and
changes the implementation to load all existing POP3 UIDs for a particular grabber into
an in-memory set and then test for membership in that set, rather than issuing a very
expensive database query for each call to shouldRetrieve().  Benchmarks indicate this as
a speedup of at least 2x.


git-svn-id: http://divmod.org/svn/Divmod/trunk@6639 866e43f7-fbfc-0310-8f2a-ec88d1da2979