Merge quotient-indexes-1054
Author: exarkun
Reviewer: glyph
Fixes #992 and #1054

To Axiom' tag functionality, this adds a more efficient way to get a list of
all the unique tags that have been applied in a Catalog.  It also adds a benchmark
for tags.

To the benchmarking system, this adds support for excluding startup and shutdown
code from the main benchmark results.  The benchmark reporter now emits two sets
of stats per run, one for the whole thing and one for the section between a start
and a stop command.

Quotient is updated to use the new functionality of Axiom for finding tag names.

All existing benchmarks have been updated to use the new functionality of the
benchmarking system as well.


git-svn-id: http://divmod.org/svn/Divmod/trunk@6959 866e43f7-fbfc-0310-8f2a-ec88d1da2979