Fix cross-store transactions, and substore attachment.
Previously generated SQL was cached on item classes, which was incorrect
because different databases could potentially be attached to their parents,
changing the SQL that was generated.  Now the SQL is properly cached per-store.

That change will also be useful to anyone wanting to implement distinct
backends for different Store objects.

Also, cross-store transactions were previously broken because items on attached
substores are not checkpointed.

Together, this change should make transactions between and across substores
work properly in the case of both success and failure.

Author: soyt

Reviewer: glyph

Fixes #1092

Fixes #1087


git-svn-id: http://divmod.org/svn/Divmod/trunk@7029 866e43f7-fbfc-0310-8f2a-ec88d1da2979