'axiomatic web --list' no longer modifies database
There was a bug where any 'axiomatic web' command would re-set the logfile to
'httpd.log' and install a bunch of objects if they weren't there already.  This
fixes it so that --list is idempotent (and also so that you won't change the
http log location if you don't want to).


Author: glyph

Reviewer: exarkun

Fixes #653 (again)


git-svn-id: http://divmod.org/svn/Divmod/trunk@7087 866e43f7-fbfc-0310-8f2a-ec88d1da2979