Decouple table names from type names, indirected through the Store.
Author: soyt

Reviewer: glyph

Fixes #400

Type names are now determined on a per-store basis, rather than cached globally
on the Item.  This is a significant improvement in correctness, allowing the
use of cross-store transactions, as well as laying the groundwork for the
inclusion of other databases which don't have SQLite's nice absence of
table-name and column-name length restrictions.


git-svn-id: http://divmod.org/svn/Divmod/trunk@7294 866e43f7-fbfc-0310-8f2a-ec88d1da2979