Minor fix to allow Vertex tests to pass on Win32 platforms.
Author: Vitorio

Reviewer: glyph

Fixes #1215

This is an extremely minor change which, while not suggested in the form of a
diff, was extremely clearly specified in the ticket.  The filenames previously
chosen were ill-advised: "TEST->TEST" is not a valid filename on Windows due to
the greater-than sign, and it is at best very confusing on UNIX.

This changes the test filename to "TESTtoTEST".


git-svn-id: http://divmod.org/svn/Divmod/trunk@7331 866e43f7-fbfc-0310-8f2a-ec88d1da2979