Merge search-1316-2
Author: exarkun
Reviewer: glyph
Fixes #1316, #1242, #1326, #1340

Factors the view and model portions of ScrollableFragment into separate classes and
adds a second model implementation which works on any sequence.  Changes the PyLucene
indexer to return a sequence which can be used here and Quotient to use this class
where appropriate.

This also changes the PyLucene indexer so that it does not store the text parts of
messages it indexes and automatically tags each document with its type.


git-svn-id: http://divmod.org/svn/Divmod/trunk@8051 866e43f7-fbfc-0310-8f2a-ec88d1da2979