Merge multi-nit-1396
Author: exarkun
Reviewer: glyph
Fixes #1396

This adds support for test method discovery to nit test cases, allowing multiple
test methods to be defined per test case.

This also changes the way reporting is done to collect all of the results in a
more compact format and move the "scratch" part of the page manipulated by the
actual tests off to the side by itself.


git-svn-id: http://divmod.org/svn/Divmod/trunk@8300 866e43f7-fbfc-0310-8f2a-ec88d1da2979